splits parsers
