+++ adapters/jwplayer/utils/contentTargeting.go
@@ -6,7 +6,6 @@ import (
 	"github.com/mxmCherry/openrtb/v15/openrtb2"
 	"net/http"
 	"net/url"
-	"strings"
 )
 
 const jwplayerSegtax = 502
@@ -38,19 +37,16 @@ type JWTargetingData struct {
 }
 
 func EnnrichRequest(request *openrtb2.BidRequest, publisherId string) {
-	println("begin enrichment")
 	if site := request.Site; site != nil {
-		AddTargeting(&site.Keywords, site.Content, publisherId)
+		Enrich(&site.Keywords, site.Content, publisherId)
 	}
 
 	if app := request.App; app != nil {
-		AddTargeting(&app.Keywords, app.Content, publisherId)
+		Enrich(&app.Keywords, app.Content, publisherId)
 	}
-
-	println("end enrichment")
 }
 
-func AddTargeting(keywords *string, content *openrtb2.Content, publisherId string) {
+func Enrich(keywords *string, content *openrtb2.Content, publisherId string) {
 	jwpsegs := GetExistingJwpsegs(content.Data)
 	if jwpsegs != nil && len(jwpsegs) > 0 {
 		writeToKeywords(keywords, jwpsegs)
@@ -75,7 +71,7 @@ func AddTargeting(keywords *string, content *openrtb2.Content, publisherId strin
 		return
 	}
 
-	jwpsegs = GetAllSegments(targetingResponse.Data)
+	jwpsegs = GetAllJwpsegs(targetingResponse.Data)
 	if len(jwpsegs) == 0 {
 		// error: segments from req were empty
 		return
@@ -83,57 +79,8 @@ func AddTargeting(keywords *string, content *openrtb2.Content, publisherId strin
 
 	writeToKeywords(keywords, jwpsegs)
 
-	contentDatum := GetContentDatum(jwpsegs)
+	contentDatum := MakeOrtbDatum(jwpsegs)
 	content.Data = append(content.Data, contentDatum)
-}
-
-func writeToKeywords(keywords *string, segments []string) {
-	if len(*keywords) > 0 {
-		*keywords += ","
-	}
-	*keywords += GetKeywords(segments)
-}
-
-func ParseContentMetadata(content openrtb2.Content) JWContentMetadata {
-	metadata := JWContentMetadata{
-		Url: content.URL,
-		Title: content.Title,
-	}
-
-	contentExt := JWContentExt{}
-	if error := json.Unmarshal(content.Ext, &contentExt); error == nil {
-		metadata.Description = contentExt.Description
-	}
-
-	return metadata
-}
-
-func GetExistingJwpsegs(data []openrtb2.Data) []string {
-	for _, datum := range data {
-		if HasJwpsegs(datum) {
-			return ParseJwpsegs(datum.Segment)
-		}
-	}
-	
-	return nil
-}
-
-func HasJwpsegs(datum openrtb2.Data) bool {
-	dataExt := JWDataExt{}
-	if error := json.Unmarshal(datum.Ext, &dataExt); error != nil {
-		return false
-	}
-	
-	return datum.Name == jwplayerDomain && dataExt.Segtax == jwplayerSegtax
-}
-
-func ParseJwpsegs(segments []openrtb2.Segment) []string {
-	jwpsegs := make([]string, len(segments))
-	for _, segment := range segments {
-		jwpsegs = append(jwpsegs, segment.Value)
-	}
-	
-	return jwpsegs
 }
 
 func FetchContentTargeting(publisherId string, contentMetadata JWContentMetadata, c chan *JWTargetingResponse) {
@@ -161,39 +108,9 @@ func FetchContentTargeting(publisherId string, contentMetadata JWContentMetadata
 	c <- &targetingResponse
 }
 
-func GetAllSegments(targeting JWTargetingData) []string {
-	return append(targeting.BaseSegments, targeting.TargetingProfiles...)
-}
-
-func GetKeywords(segments []string) string {
-	if len(segments) == 0 {
-		return ""
+func writeToKeywords(keywords *string, jwpsegs []string) {
+	if len(*keywords) > 0 {
+		*keywords += ","
 	}
-
-	keyword := "jwpseg="
-	// expected format: jwpseg=1,jwpseg=2,jwpseg=3
-	keyword += strings.Join(segments, "," + keyword)
-	return keyword
-}
-
-func GetContentDatum(segments []string) (contentData openrtb2.Data) {
-	contentData.Name = jwplayerDomain
-	contentData.Segment = GetContentSegments(segments)
-	dataExt := JWDataExt{
-		Segtax: jwplayerSegtax,
-	}
-	contentData.Ext, _ = json.Marshal(dataExt)
-	return contentData
-}
-
-func GetContentSegments(rawSegments []string) []openrtb2.Segment {
-	segments := make([]openrtb2.Segment, len(rawSegments))
-	for _, rawSegment := range rawSegments {
-		segment := openrtb2.Segment{
-			Value: rawSegment,
-		}
-		segments = append(segments, segment)
-	}
-
-	return segments
+	*keywords += GetXandrKeywords(jwpsegs)
 }
